pyproject.toml
[tool.poetry]
name = "poetry"
version = "0.1.0"
description = ""
authors = ["Abigail <abigail.cabascango@ioet.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
pytest = "^8.2.2"
pydantic = "^2.7.3"
mypy = "^1.10.0"
uvicorn = "^0.30.1"
flake8 = "^7.0.0"
pytest-cov = "^5.0.0"
pytest-mock = "^3.14.0"
faker = "^25.8.0"
pyjwt = "^2.8.0"
passlib = "^1.7.4"
boto3 = "^1.34.133"
black = "^24.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
